Updates base image to phusion/focal-1.0.0
This image is based on Ubuntu 18.04 (previous was
Ubuntu 14.04). Previously, the reqs image was not
buildable - I believe this is due to the older
Python experiencing SSL errors on pypi.org (though
to be honest, I didn't spent a lot of time finding
out the *exact* cause).

This also manually installs some dependencies, since
running `apt-get autoremove python-dev` will actually
remove `libpython2.7`.

As far as I can tell, everything is functioning as
designed after this update.