fix for #51
some pages are missing the featuredtable div, this could
cause an exception since soup_featured_table may be a `NoneType`
with no attribute named `table`.

I'm not sure if any artist pages are missing the `discotable` as well,
since the logic is about the same I applied a pre-emptive fix to it
as well.